refactor: move ScanContext and ScanMode to ggshield.scan
