+++ ggshield/cmd/secret/scan/ci.py
@@ -5,13 +5,8 @@ import click
 
 from ggshield.core.cache import ReadOnlyCache
 from ggshield.core.git_shell import check_git_dir, get_list_commit_SHA
-from ggshield.core.utils import (
-    EMPTY_SHA,
-    ScanContext,
-    ScanMode,
-    SupportedCI,
-    handle_exception,
-)
+from ggshield.core.utils import EMPTY_SHA, SupportedCI, handle_exception
+from ggshield.scan import ScanContext, ScanMode
 from ggshield.scan.repo import scan_commit_range
