+++ ggshield/cmd/secret/scan/prereceive.py
@@ -10,14 +10,9 @@ import click
 from ggshield.core.cache import ReadOnlyCache
 from ggshield.core.git_shell import get_list_commit_SHA, git
 from ggshield.core.text_utils import display_error
-from ggshield.core.utils import (
-    EMPTY_SHA,
-    PRERECEIVE_TIMEOUT,
-    ScanContext,
-    ScanMode,
-    handle_exception,
-)
+from ggshield.core.utils import EMPTY_SHA, PRERECEIVE_TIMEOUT, handle_exception
 from ggshield.output import GitLabWebUIOutputHandler
+from ggshield.scan import ScanContext, ScanMode
 from ggshield.scan.repo import scan_commit_range
