+++ ggshield/scan/__init__.py
@@ -1,3 +1,5 @@
+from .scan_context import ScanContext
+from .scan_mode import ScanMode
 from .scannable import Commit, File, Files
 from .scanner import Result, Results, ScanCollection, SecretScanner
 
@@ -10,4 +12,6 @@ __all__ = [
     "Results",
     "SecretScanner",
     "ScanCollection",
+    "ScanContext",
+    "ScanMode",
 ]
