+++ tests/unit/cmd/scan/test_prepush.py
@@ -6,13 +6,8 @@ from click.testing import CliRunner
 
 from ggshield.cmd.main import cli
 from ggshield.core.filter import init_exclusion_regexes
-from ggshield.core.utils import (
-    EMPTY_SHA,
-    EMPTY_TREE,
-    IGNORED_DEFAULT_WILDCARDS,
-    ScanContext,
-    ScanMode,
-)
+from ggshield.core.utils import EMPTY_SHA, EMPTY_TREE, IGNORED_DEFAULT_WILDCARDS
+from ggshield.scan import ScanContext, ScanMode
 from ggshield.scan.repo import cd
 from tests.repository import Repository
 from tests.unit.conftest import assert_invoke_ok
