+++ tests/unit/output/test_json_output.py
@@ -4,10 +4,10 @@ import pytest
 from pytest_voluptuous import Partial, S
 from voluptuous import Optional, validators
 
-from ggshield.core.utils import Filemode, ScanContext, ScanMode
+from ggshield.core.utils import Filemode
 from ggshield.output import JSONOutputHandler, OutputHandler
 from ggshield.output.json.schemas import JSONScanCollectionSchema
-from ggshield.scan import Commit, ScanCollection, SecretScanner
+from ggshield.scan import Commit, ScanCollection, ScanContext, ScanMode, SecretScanner
 from tests.unit.conftest import (
     _MULTIPLE_SECRETS_PATCH,
     _NO_SECRET_PATCH,
